Initial version of "atlassian-sdk" image
This installs the latest version available from atlassian debian repository, which at this moment
is "6.1.0". It uses sdk version 6.1.0. It also adds an environment variable that points to the
installation directory of the SDK.

Added links to each of the available images in README file

Added ATLAS_HOME environment variable that points to SDK installation dir

Revert "Initial commit"

This reverts commit 1e0f8788c6c7d26fc07cf8fe3005f5e5d1159254, as these files belong to
the "oracle-jdk8" image.